Proyecto-Final/requirements.txt
Flask>=2.0
pandas
numpy
plotly
sqlalchemy
pyodbc
python-dotenv
scipy
faker
